+++ pkg/scan/local_scan.go
@@ -2,60 +2,16 @@ package scan
 
 import (
 	"archive/tar"
-	"bytes"
 	"context"
-	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"io"
 	"os"
-	"path"
 	"slices"
-	"strings"
-
-	"github.com/anchore/syft/syft/format"
-	"github.com/anchore/syft/syft/format/cyclonedxjson"
-	v1 "github.com/google/go-containerregistry/pkg/v1"
-	"github.com/klauspost/compress/zstd"
-	"go.uber.org/zap"
 
 	"github.com/defenseunicorns/uds-security-hub/internal/executor"
 	"github.com/defenseunicorns/uds-security-hub/pkg/types"
 )
-
-const (
-	SbomFilename = "sboms.tar"
-)
-
-type imageRef interface {
-	TrivyCommand() []string
-}
-
-type remoteImageRef struct {
-	ImageRef string
-}
-
-func (r *remoteImageRef) TrivyCommand() []string {
-	return []string{"image", "--image-src=remote", r.ImageRef}
-}
-
-type cyclonedxSbomRef struct {
-	ArtifactName string
-	SBOMFile     string
-}
-
-func (s *cyclonedxSbomRef) TrivyCommand() []string {
-	return []string{"sbom", s.SBOMFile}
-}
-
-type rootfsRef struct {
-	ArtifactName string
-	RootFSDir    string
-}
-
-func (r rootfsRef) TrivyCommand() []string {
-	return []string{"rootfs", r.RootFSDir}
-}
 
 func extractFilesFromTar(r io.Reader, filenames ...string) (map[string][]byte, error) {
 	tarReader := tar.NewReader(r)
@@ -81,284 +37,6 @@ func extractFilesFromTar(r io.Reader, filenames ...string) (map[string][]byte, e
 	}
 
 	return results, nil
-}
-
-func extractSBOMTarFromZarfPackage(tarFilePath string) ([]byte, error) {
-	file, err := os.Open(tarFilePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open tar file: %w", err)
-	}
-	defer file.Close()
-
-	zstdReader, err := zstd.NewReader(file)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create zstd reader: %w", err)
-	}
-	defer zstdReader.Close()
-	files, err := extractFilesFromTar(zstdReader, SbomFilename)
-	if err != nil {
-		return nil, err
-	}
-
-	return files[SbomFilename], nil
-}
-
-func extractArtifactInformationFromSBOM(r io.Reader) string {
-	type SyftSbomHeader struct {
-		Source struct {
-			Metadata struct {
-				Tags []string `json:"tags"`
-			} `json:"metadata"`
-		} `json:"source"`
-	}
-
-	var sbomHeader SyftSbomHeader
-
-	err := json.NewDecoder(r).Decode(&sbomHeader)
-	if err != nil {
-		return ""
-	}
-
-	if len(sbomHeader.Source.Metadata.Tags) == 0 {
-		return ""
-	}
-
-	return sbomHeader.Source.Metadata.Tags[0]
-}
-
-func convertToCyclonedxFormat(header *tar.Header, r io.Reader, outputDir string) (*cyclonedxSbomRef, error) {
-	cyclonedxEncoder, err := cyclonedxjson.NewFormatEncoderWithConfig(cyclonedxjson.DefaultEncoderConfig())
-	if err != nil {
-		return nil, fmt.Errorf("failed to create cyclonedx encoder: %w", err)
-	}
-
-	sbomData, err := io.ReadAll(r)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read sbom from tar for %q: %w", header.Name, err)
-	}
-
-	artifactName := extractArtifactInformationFromSBOM(bytes.NewReader(sbomData))
-	if artifactName == "" {
-		// default to the filename if we were unable to extract anything meaningful
-		artifactName = header.Name
-	}
-
-	sbom, _, _, err := format.Decode(bytes.NewReader(sbomData))
-	if err != nil {
-		return nil, fmt.Errorf("failed to convert sbom format for %q: %w", header.Name, err)
-	}
-
-	cyclonedxBytes, err := format.Encode(*sbom, cyclonedxEncoder)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encode cyclonnedx format for %q: %w", header.Name, err)
-	}
-
-	// use a sha256 for the filename in the tar to avoid any security issues with malformed tar
-	sbomSha256 := sha256.Sum256(cyclonedxBytes)
-	cyclonedxSBOMFilename := path.Join(outputDir, fmt.Sprintf("%x", sbomSha256))
-	if err := os.WriteFile(cyclonedxSBOMFilename, cyclonedxBytes, header.FileInfo().Mode().Perm()); err != nil {
-		return nil, fmt.Errorf("failed to write new cyclonnedx file for %q: %w", header.Name, err)
-	}
-
-	return &cyclonedxSbomRef{
-		ArtifactName: artifactName,
-		SBOMFile:     cyclonedxSBOMFilename,
-	}, nil
-}
-
-func extractSBOMImageRefsFromReader(r io.Reader) ([]*cyclonedxSbomRef, error) {
-	tmp, err := os.MkdirTemp("", "zarf-sbom-spdx-files-*")
-	if err != nil {
-		return nil, fmt.Errorf("failed to create tmp dir: %w", err)
-	}
-
-	var results []*cyclonedxSbomRef
-
-	sbomTarReader := tar.NewReader(r)
-	for {
-		header, err := sbomTarReader.Next()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return nil, fmt.Errorf("failed to read header in sbom tar: %w", err)
-		}
-
-		if strings.HasSuffix(header.Name, ".json") {
-			sbomImageRef, err := convertToCyclonedxFormat(header, sbomTarReader, tmp)
-			if err != nil {
-				return nil, err
-			}
-			results = append(results, sbomImageRef)
-		}
-	}
-
-	return results, nil
-}
-
-func scannableImage(name string) bool {
-	if strings.HasSuffix(name, ".att") || strings.HasSuffix(name, ".sig") {
-		return false
-	}
-
-	return true
-}
-
-func replacePathChars(s string) string {
-	s = strings.ReplaceAll(s, "/", "-")
-	s = strings.ReplaceAll(s, ":", "_")
-	return s
-}
-
-type ExtractRootFSResult struct {
-	RootPath string
-	Refs     []rootfsRef
-}
-
-func (e *ExtractRootFSResult) Cleanup() error {
-	return os.RemoveAll(e.RootPath)
-}
-
-func ExtractRootFS(logger types.Logger, tarFilePath string, command types.CommandExecutor) (*ExtractRootFSResult, error) {
-	f, err := os.Open(tarFilePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open tar: %w", err)
-	}
-
-	r, err := zstd.NewReader(f)
-	if err != nil {
-		return nil, fmt.Errorf("failed to unzstd tar: %w", err)
-	}
-
-	tarBytes, err := io.ReadAll(r)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read tar: %w", err)
-	}
-
-	files, err := extractFilesFromTar(bytes.NewReader(tarBytes), "images/index.json")
-	if err != nil {
-		return nil, fmt.Errorf("failed to extract index.json: %w", err)
-	}
-
-	var imagesIndex v1.IndexManifest
-	if err := json.Unmarshal(files["images/index.json"], &imagesIndex); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal index.json: %w", err)
-	}
-
-	imageNameToManifestFile := make(map[string]string)
-	var manifestFilesToExtract []string
-	for i := range imagesIndex.Manifests {
-		digest := imagesIndex.Manifests[i].Digest.Hex
-		name := imagesIndex.Manifests[i].Annotations["org.opencontainers.image.base.name"]
-
-		if !scannableImage(name) {
-			continue
-		}
-
-		layerLocation := fmt.Sprintf("images/blobs/sha256/%s", digest)
-		imageNameToManifestFile[name] = layerLocation
-		manifestFilesToExtract = append(manifestFilesToExtract, layerLocation)
-	}
-
-	extractedManifests, err := extractFilesFromTar(bytes.NewReader(tarBytes), manifestFilesToExtract...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to extract image manifests: %w", err)
-	}
-
-	imageNameToParsedManifest := make(map[string]v1.Manifest)
-	for imageName, manifestFileName := range imageNameToManifestFile {
-		var packagedManifest v1.Manifest
-		err := json.Unmarshal(extractedManifests[manifestFileName], &packagedManifest)
-		if err != nil {
-			return nil, fmt.Errorf("failed to unmarshal image manifest: %w", err)
-		}
-		imageNameToParsedManifest[imageName] = packagedManifest
-	}
-
-	tmpDir, err := os.MkdirTemp("", "uds-local-scan-*")
-	if err != nil {
-		return nil, fmt.Errorf("failed to create tmp dir: %w", err)
-	}
-
-	tarPkgOnDisk := path.Join(tmpDir, "pkg.tar")
-	if err := os.WriteFile(tarPkgOnDisk, tarBytes, 0o600); err != nil {
-		return nil, fmt.Errorf("failed to write tar file: %w", err)
-	}
-
-	pkgOutDir := path.Join(tmpDir, "pkg")
-	if err := os.Mkdir(pkgOutDir, 0o700); err != nil {
-		return nil, fmt.Errorf("failed to create output pkg dir: %w", err)
-	}
-	_, _, err = command.ExecuteCommand(
-		"tar",
-		[]string{"-xf", tarPkgOnDisk, "-C", pkgOutDir},
-		nil,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to untar package: %w", err)
-	}
-
-	var results []rootfsRef
-
-	for imageName, manifest := range imageNameToParsedManifest {
-		imageRootFS := path.Join(tmpDir, replacePathChars(imageName))
-		if err := os.Mkdir(imageRootFS, 0o700); err != nil {
-			return nil, fmt.Errorf("failed to create dir for image %s: %w", imageName, err)
-		}
-		for i := range manifest.Layers {
-			digest := manifest.Layers[i].Digest.Hex
-			layerBlob := path.Join(pkgOutDir, "images", "blobs", "sha256", digest)
-			_, stderr, err := command.ExecuteCommand(
-				"tar",
-				[]string{"-zvxf", layerBlob, "-C", imageRootFS},
-				nil,
-			)
-			if err != nil {
-				logger.Warn(
-					"error occurred while untar-ing layer",
-					zap.String("imageName", imageName),
-					zap.String("digest", digest),
-					zap.String("stderr", stderr),
-					zap.Error(err),
-				)
-			}
-		}
-		results = append(results, rootfsRef{
-			ArtifactName: imageName,
-			RootFSDir:    imageRootFS,
-		})
-
-		// add read and write permissions to the user or the trivy scan and the cleanup will fail
-		_, _, err = command.ExecuteCommand(
-			"chmod",
-			[]string{"-R", "u+rw", imageRootFS},
-			nil,
-		)
-		if err != nil {
-			logger.Warn(
-				"unable to ensure proper permissions on extracted files",
-				zap.String("imageName", imageName),
-				zap.Error(err),
-			)
-		}
-	}
-
-	return &ExtractRootFSResult{RootPath: tmpDir, Refs: results}, nil
-}
-
-// ExtractSBOMsFromTar extracts images from the tar archive and returns names of the container images.
-// Parameters:
-// - tarFilePath: the path to the tar archive to extract the images from.
-// Returns:
-// - []sbomImageRef: references to images and their sboms.
-// - error: an error if the extraction fails.
-func ExtractSBOMsFromTar(tarFilePath string) ([]*cyclonedxSbomRef, error) {
-	sbomTar, err := extractSBOMTarFromZarfPackage(tarFilePath)
-	if err != nil {
-		return nil, err
-	}
-
-	return extractSBOMImageRefsFromReader(bytes.NewReader(sbomTar))
 }
 
 // LocalPackageScanner is a struct that holds the logger and paths for docker configuration and package.
