+++ pkg/scan/sbom_test.go
@@ -1,0 +1,37 @@
+package scan
+
+import "testing"
+
+func TestExtractSBOMsFromTar(t *testing.T) {
+	filePath := "testdata/zarf-package-mattermost-arm64-9.9.1-uds.0.tar.zst"
+	refs, err := ExtractSBOMsFromTar(filePath)
+	if err != nil {
+		t.Fatalf("Failed to extract images from tar: %v", err)
+	}
+
+	if len(refs) == 0 {
+		t.Fatal("Expected non-empty images, got empty")
+	}
+
+	expectedImageNameFromSBOM := []string{
+		"docker.io/appropriate/curl:latest",
+	}
+
+	for _, sbomName := range expectedImageNameFromSBOM {
+		found := false
+		for _, ref := range refs {
+			if ref.ArtifactName == sbomName {
+				found = true
+				t.Logf("Found expected image: %s", sbomName)
+
+				if ref.SBOMFile == "" {
+					t.Error("got an empty sbomfile, this will not be scannable by trivy")
+				}
+				break
+			}
+		}
+		if !found {
+			t.Errorf("Expected image not found: %s", sbomName)
+		}
+	}
+}
